move conftest
