--- package/tests/test_integrations/confest.py
@@ -1,35 +1,0 @@
-"""
-This file contains the fixtures that are reusable by any tests within
-this directory. You don't need to import the fixtures as pytest will
-discover them automatically. More info here:
-https://docs.pytest.org/en/latest/fixture.html
-"""
-
-from pytest import fixture
-
-from kedro.io.core import generate_timestamp
-
-
-@fixture(params=[None])
-def load_version(request):
-    return request.param
-
-
-@fixture(params=[None])
-def save_version(request):
-    return request.param or generate_timestamp()
-
-
-@fixture(params=[None])
-def load_args(request):
-    return request.param
-
-
-@fixture(params=[None])
-def save_args(request):
-    return request.param
-
-
-@fixture(params=[None])
-def fs_args(request):
-    return request.param
